[976] Refactor directory structure and file locations
