--- projects/_03_independent_cqc/_04a_preprocessing/fargate/README.md
@@ -1,3 +1,0 @@
-# Preprocessing
-
-The preprocessing task carries out essential cleaning of feature data to ensure suitability for use in model training.
