+++ projects/_03_independent_cqc/_05_model/fargate/retraining/Dockerfile
